refactor: extract chmodRecursive into file permission service
